+++ familyhelper-webapi-node/src/main/java/com/dwarfeng/familyhelper/webapi/node/controller/v1/project/MemoRemindDriverSupportController.java
@@ -1,0 +1,102 @@
+package com.dwarfeng.familyhelper.webapi.node.controller.v1.project;
+
+import com.dwarfeng.familyhelper.project.sdk.bean.entity.FastJsonMemoRemindDriverSupport;
+import com.dwarfeng.familyhelper.project.stack.bean.entity.MemoRemindDriverSupport;
+import com.dwarfeng.familyhelper.webapi.stack.service.project.MemoRemindDriverSupportResponseService;
+import com.dwarfeng.subgrade.sdk.bean.dto.FastJsonResponseData;
+import com.dwarfeng.subgrade.sdk.bean.dto.JSFixedFastJsonPagedData;
+import com.dwarfeng.subgrade.sdk.bean.dto.PagingUtil;
+import com.dwarfeng.subgrade.sdk.bean.dto.ResponseDataUtil;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.SkipRecord;
+import com.dwarfeng.subgrade.sdk.interceptor.login.LoginRequired;
+import com.dwarfeng.subgrade.stack.bean.BeanTransformer;
+import com.dwarfeng.subgrade.stack.bean.dto.PagedData;
+import com.dwarfeng.subgrade.stack.bean.dto.PagingInfo;
+import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
+import com.dwarfeng.subgrade.stack.exception.ServiceExceptionMapper;
+import org.springframework.web.bind.annotation.*;
+
+import javax.servlet.http.HttpServletRequest;
+
+/**
+ * 备忘录提醒驱动器支持控制器。
+ *
+ * @author DwArFeng
+ * @since 1.0.9
+ */
+@RestController("projectMemoRemindDriverSupportController")
+@RequestMapping("/api/v1/project")
+public class MemoRemindDriverSupportController {
+
+    private final MemoRemindDriverSupportResponseService service;
+    private final ServiceExceptionMapper sem;
+
+    private final BeanTransformer<MemoRemindDriverSupport, FastJsonMemoRemindDriverSupport> beanTransformer;
+
+    public MemoRemindDriverSupportController(
+            MemoRemindDriverSupportResponseService service, ServiceExceptionMapper sem,
+            BeanTransformer<MemoRemindDriverSupport, FastJsonMemoRemindDriverSupport> beanTransformer
+    ) {
+        this.service = service;
+        this.sem = sem;
+        this.beanTransformer = beanTransformer;
+    }
+
+    @GetMapping("/memo-remind-driver-support/{id}/exists")
+    @BehaviorAnalyse
+    @LoginRequired
+    public FastJsonResponseData<Boolean> exists(HttpServletRequest request, @PathVariable("id") String id) {
+        try {
+            boolean exists = service.exists(new StringIdKey(id));
+            return FastJsonResponseData.of(ResponseDataUtil.good(exists));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(e, sem));
+        }
+    }
+
+    @GetMapping("/memo-remind-driver-support/{id}")
+    @BehaviorAnalyse
+    @LoginRequired
+    public FastJsonResponseData<FastJsonMemoRemindDriverSupport> get(HttpServletRequest request, @PathVariable("id") String id) {
+        try {
+            MemoRemindDriverSupport memoRemindDriverSupport = service.get(new StringIdKey(id));
+            return FastJsonResponseData.of(ResponseDataUtil.good(FastJsonMemoRemindDriverSupport.of(memoRemindDriverSupport)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(e, sem));
+        }
+    }
+
+    @GetMapping("/memo-remind-driver-support/all")
+    @BehaviorAnalyse
+    @SkipRecord
+    @LoginRequired
+    public FastJsonResponseData<JSFixedFastJsonPagedData<FastJsonMemoRemindDriverSupport>> all(
+            HttpServletRequest request, @RequestParam("page") int page, @RequestParam("rows") int rows
+    ) {
+        try {
+            PagedData<MemoRemindDriverSupport> all = service.all(new PagingInfo(page, rows));
+            PagedData<FastJsonMemoRemindDriverSupport> transform = PagingUtil.transform(all, beanTransformer);
+            return FastJsonResponseData.of(ResponseDataUtil.good(JSFixedFastJsonPagedData.of(transform)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(e, sem));
+        }
+    }
+
+    @GetMapping("/memo-remind-driver-support/id-like")
+    @BehaviorAnalyse
+    @SkipRecord
+    @LoginRequired
+    public FastJsonResponseData<JSFixedFastJsonPagedData<FastJsonMemoRemindDriverSupport>> idLike(
+            HttpServletRequest request,
+            @RequestParam("pattern") String pattern, @RequestParam("page") int page, @RequestParam("rows") int rows
+    ) {
+        try {
+            PagedData<MemoRemindDriverSupport> all = service.idLike(pattern, new PagingInfo(page, rows));
+            PagedData<FastJsonMemoRemindDriverSupport> transform = PagingUtil.transform(all, beanTransformer);
+            return FastJsonResponseData.of(ResponseDataUtil.good(JSFixedFastJsonPagedData.of(transform)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(e, sem));
+        }
+    }
+}