+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 #### 功能构建
 
+- 新建 Controller。
+  - com.dwarfeng.familyhelper.webapi.node.controller.v1.project.MemoRemindDriverInfoController。
+  - com.dwarfeng.familyhelper.webapi.node.controller.v1.project.MemoRemindDriverSupportController。
+
 - 优化响应服务的行为。
   - com.dwarfeng.familyhelper.webapi.impl.service.system.PexpResponseServiceImpl。
   - com.dwarfeng.familyhelper.webapi.impl.service.system.RoleResponseServiceImpl。